+++ src/driver/bin/odoc_driver_monorepo.ml
@@ -50,7 +50,8 @@ let real_run ~odoc_dir ~odocl_dir ~index_dir ~mld_dir path extra_pkgs extra_libs
         Compile.init_stats units;
         let compiled = Compile.compile ~partial_dir:odoc_dir units in
         let linked =
-          Compile.link ~warnings_tags:[] ~custom_layout:true compiled
+          Compile.link ~partial_dir:odoc_dir ~warnings_tags:[]
+            ~custom_layout:true compiled
         in
         let occurrence_file =
           let output =